server/requirements/common.txt
Django==2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
gunicorn==19.9.0
drf-yasg==1.17.0

